backend/requirements.txt
fastapi[all]
uvicorn
sqlalchemy
pydantic
passlib[bcrypt]
python-jose
requests
indent
dependency_injector
huggingface_hub
transformers
tweepy
python-dotenv
python-multipart
rich

pytest
psutil
diffusers
torch

